+++ src/tcpedit/plugins/dlt_hdlc/hdlc.h
@@ -29,14 +29,16 @@ int dlt_hdlc_register(tcpeditdlt_t *ctx);
 int dlt_hdlc_init(tcpeditdlt_t *ctx);
 int dlt_hdlc_cleanup(tcpeditdlt_t *ctx);
 int dlt_hdlc_parse_opts(tcpeditdlt_t *ctx);
-int dlt_hdlc_decode(tcpeditdlt_t *ctx, const u_char *packet, const int pktlen);
-int dlt_hdlc_encode(tcpeditdlt_t *ctx, u_char *packet, int pktlen, tcpr_dir_t dir);
-int dlt_hdlc_proto(tcpeditdlt_t *ctx, const u_char *packet, const int pktlen);
-u_char *dlt_hdlc_get_layer3(tcpeditdlt_t *ctx, u_char *packet, const int pktlen);
-u_char *dlt_hdlc_merge_layer3(tcpeditdlt_t *ctx, u_char *packet, const int pktlen, u_char *l3data);
+int dlt_hdlc_decode(tcpeditdlt_t *ctx, const u_char *packet, const size_t pktlen);
+int dlt_hdlc_encode(tcpeditdlt_t *ctx, u_char *packet, size_t pktlen, tcpr_dir_t dir);
+int dlt_hdlc_proto(tcpeditdlt_t *ctx, const u_char *packet, const size_t pktlen);
+u_char *dlt_hdlc_get_layer3(tcpeditdlt_t *ctx, u_char *packet, const size_t pktlen);
+u_char *dlt_hdlc_merge_layer3(tcpeditdlt_t *ctx, u_char *packet,
+        const size_t pktlen, u_char *l3data);
 tcpeditdlt_l2addr_type_t dlt_hdlc_l2addr_type(void);
-int dlt_hdlc_l2len(tcpeditdlt_t *ctx, const u_char *packet, const int pktlen);
-u_char *dlt_hdlc_get_mac(tcpeditdlt_t *ctx, tcpeditdlt_mac_type_t mac, const u_char *packet, const int pktlen);
+int dlt_hdlc_l2len(tcpeditdlt_t *ctx, const u_char *packet, const size_t pktlen);
+u_char *dlt_hdlc_get_mac(tcpeditdlt_t *ctx, tcpeditdlt_mac_type_t mac,
+        const u_char *packet, const size_t pktlen);
 
 #endif
